% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_excerpt_by_district.R
\name{batch_excerpt_by_district}
\alias{batch_excerpt_by_district}
\title{Internal helper for extracting a single district's data from WISEDash files}
\usage{
batch_excerpt_by_district(.input_path, .pattern, .district_code, .output_path)
}
\arguments{
\item{.input_path}{the location of the source files}

\item{.pattern}{a glob to find the files}

\item{.district_code}{a pattern to identify a district's data}

\item{.output_path}{the location to write the resulting files}
}
\description{
Internal helper for extracting a single district's data from WISEDash files
}
\keyword{internal}
