% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.R
\name{count_if}
\alias{count_if}
\title{Helper for conditionally including values in a total}
\usage{
count_if(.counts, .mask)
}
\arguments{
\item{.counts}{\emph{<int>} counts, a la \code{STUDENT_COUNT}, that may be incorporated in a total}

\item{.mask}{\emph{<lgl>} whether or not a value of \code{.counts} should be included}
}
\value{
\code{.counts} when \code{.mask} is true, or zero otherwise
}
\description{
Helper for conditionally including values in a total
}
\keyword{internal}
