% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hercwidpi-package.R
\docType{package}
\name{hercwidpi-package}
\alias{hercwidpi}
\alias{hercwidpi-package}
\title{hercwidpi: utilities and specifications for parsing data from the WI DPI}
\description{
Work with files downloaded from the Wisconsin Department of Public Instruction. The basic action is to parse annual text files and wrangle them into tidy multi-year tables. There are utilities for summarizing -- by area, agency, or demographic grouping. Additional tools use percentiles to compare across otherwise incompatible test results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://higherx4racine.github.io/hercwidpi/}
}

}
\author{
\strong{Maintainer}: Ben Taft \email{ben.taft@career2cradle.org}

Other contributors:
\itemize{
  \item Higher Expectations for Racine County [copyright holder]
}

}
\keyword{internal}
