% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.R
\name{is_proficient}
\alias{is_proficient}
\title{Helper predicate for detecting when students passed the Forward exam.}
\usage{
is_proficient(.result_codes, .proficiency_threshold = 3L)
}
\arguments{
\item{.result_codes}{\emph{<chr>} usually the TEST_RESULT_CODE column from a WISEDash data set.}

\item{.proficiency_threshold}{\emph{<int>} the lowest result code that counts as success}
}
\value{
\emph{<lgl>} \code{TRUE} when the numeric value of \code{.result_codes} meets or exceeds the \code{.proficiency_threshold}
}
\description{
Helper predicate for detecting when students passed the Forward exam.
}
