% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.R
\name{is_tested}
\alias{is_tested}
\title{Helper predicate for detecting when students took the Forward exam.}
\usage{
is_tested(.result_codes)
}
\arguments{
\item{.result_codes}{\emph{<chr>} usually the TEST_RESULT_CODE column from a WISEDash data set.}
}
\value{
\emph{<lgl>} \code{TRUE} when the \code{.result_codes} value contains a digit
}
\description{
Helper predicate for detecting when students took the Forward exam.
}
