% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.R
\name{load_forward}
\alias{load_forward}
\title{Syntactic sugar for loading reports about the Forward Exam.}
\usage{
load_forward(.files)
}
\arguments{
\item{.files}{paths to csv files that contain data from WISEDash public}
}
\value{
a data frame with 20 columns, as described by \link{SPEC_FOR_FORWARD}
}
\description{
Syntactic sugar for loading reports about the Forward Exam.
}
\seealso{
\code{\link[=read_wisedash_public]{read_wisedash_public()}}

\link{SPEC_FOR_FORWARD}
}
