% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_completion.R
\name{load_ps_enrollment}
\alias{load_ps_enrollment}
\title{Syntactic sugar for loading reports about enrollment after high school completion.}
\usage{
load_ps_enrollment(.files)
}
\arguments{
\item{.files}{paths to csv files that contain data from WISEDash public}
}
\value{
a data frame with 20 columns, as described by \link{SPEC_FOR_PS_ENROLLMENT}
}
\description{
Syntactic sugar for loading reports about enrollment after high school completion.
}
\seealso{
\code{\link[=read_wisedash_public]{read_wisedash_public()}}

\link{SPEC_FOR_PS_ENROLLMENT}
}
