% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_completion.R
\name{load_public_hs_completion}
\alias{load_public_hs_completion}
\title{Syntactic sugar for loading reports about completing public high school}
\usage{
load_public_hs_completion(.files)
}
\arguments{
\item{.files}{paths to csv files that contain data from WISEDash public}
}
\value{
a data frame with 17 columns, as described by the specs in "See Also"
}
\description{
Syntactic sugar for loading reports about completing public high school
}
\seealso{
\code{\link[=read_wisedash_public]{read_wisedash_public()}}

\link{SPEC_FOR_HS_COMPLETION}

\link{SPEC_FOR_PUBLIC_HS_COMPLETION}
}
