% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SPEC_FOR_AGENCY_INFO}
\alias{SPEC_FOR_AGENCY_INFO}
\title{Column specification for data about districts and school buildings}
\format{
\subsection{\code{SPEC_FOR_AGENCY_INFO}}{

A list of 25 elements
\describe{
\item{SCHOOL_YEAR}{\verb{<chr>} the school year, in YYYY-YY format}
\item{CESA}{\verb{<chr>} which CESA, if any, the school belongs to}
\item{COUNTY_CODE}{\verb{<int>} A non-FIPS ID number for the county}
\item{COUNTY_NAME}{\verb{<chr>} The name of the county}
\item{ATHLETIC_CONFERENCE_CODE}{\verb{<int>} a unique identifier for the agency's conference}
\item{ATHLETIC_CONFERENCE_NAME}{\verb{<chr>} the name of the agency's conference}
\item{WTCS_CODE}{\verb{<chr>} a two-letter code that I know nothing about}
\item{LOCALE_CODE}{\verb{<int>} a two-digit code that I know nothing about}
\item{LOCALE_NAME}{\verb{<chr>}a geographic categorization like "Urban" or "Rural"}
\item{LATITUDE}{\verb{<dbl>} distance from the equator}
\item{LONGITUDE}{\verb{<dbl>} distance from Greenwich, England}
\item{FULL_NCES_CODE}{\verb{<chr>} a thirteen-digit, FIPS-like code}
\item{DISTRICT_CODE}{\verb{<chr>} the four-number code for each district, or "[Statewide]"}
\item{SCHOOL_CODE}{\verb{<chr>} the six-digit code for a school, or "[Districtwide]"}
\item{DISTRICT_NAME}{\verb{<chr>} a human-readable name for a district. Not consistent across years}
\item{AGENCY_TYPE}{\verb{<chr>} helps distinguish between district-run and independent charter schools}
\item{SCHOOL_NAME}{\verb{<chr>} a human-readable name for a school. Not consistent across years}
\item{GRADE_GROUP}{\verb{<chr>} elementary, middle, high, k-12, etc.}
\item{LOW_GRADE}{\verb{<chr>} the earliest grade in this school or district}
\item{HIGH_GRADE}{\verb{<chr>} the latest grade in this school or district}
\item{SCHOOL_SIZE}{\verb{<chr>} qualitative, from Tiny to Large, including Unknown}
\item{CHARTER_IND}{\verb{<chr>} "No" or "Yes"}
\item{VIRTUAL_SCH_IND}{\verb{<chr>} "No" or "Yes"}
\item{TITLE_1A_PROGRAM_CODE}{\verb{<chr>} a 3- or 4-character code about Title 1 status}
\item{TITLE_1A_PROGRAM_NAME}{\verb{<chr>} a human-readable description of Title 1 status}
}
}
}
\source{
\url{https://dpi.wi.gov/wisedash/public/download-files?field_wisedash_upload_type_value=Agency}
}
\usage{
SPEC_FOR_AGENCY_INFO
}
\description{
Column specification for data about districts and school buildings
}
\keyword{datasets}
