% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SPEC_FOR_HS_COMPLETION}
\alias{SPEC_FOR_HS_COMPLETION}
\title{Column specification for data about High School Completion}
\format{
\subsection{SPEC_FOR_HS_COMPLETION}{

A list of 17 elements
\describe{
\item{SCHOOL_YEAR}{\emph{<chr>} the school year, in YYYY-YY format}
\item{AGENCY_TYPE}{\emph{<chr>} helps distinguish between district-run and independent charter schools}
\item{CESA}{\emph{<chr>} which CESA, if any, the school belongs to}
\item{COUNTY}{\emph{<chr>} the county that the school is located in}
\item{DISTRICT_CODE}{\emph{<chr>} the four-number code for each district, or "[Statewide]"}
\item{SCHOOL_CODE}{\emph{<chr>} the six-digit code for a school, or "[Districtwide]"}
\item{GRADE_GROUP}{\emph{<chr>} elementary, middle, high, k-12, etc.}
\item{DISTRICT_NAME}{\emph{<chr>} a human-readable name for a district. Not consistent across years}
\item{SCHOOL_NAME}{\emph{<chr>} a human-readable name for a school. Not consistent across years}
\item{GROUP_BY}{\emph{<chr>} demographic categories like race or gender}
\item{GROUP_BY_VALUE}{\emph{<chr>} demographic groups like Asian or Trans}
\item{STUDENT_COUNT}{\emph{<int>} the number of students who completed high school}
}
}
}
\source{
\url{https://dpi.wi.gov/wisedash/download-files/type?field_wisedash_upload_type_value=hs-completion}
}
\usage{
SPEC_FOR_HS_COMPLETION
}
\description{
Column specification for data about High School Completion
}
\keyword{datasets}
