% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SPEC_FOR_PRIVATE_HS_COMPLETION}
\alias{SPEC_FOR_PRIVATE_HS_COMPLETION}
\title{Column specification for data about Completion at Private High Schools}
\format{
\subsection{SPEC_FOR_PRIVATE_HS_COMPLETION}{

A list of 17 elements
\describe{
\item{SCHOOL_YEAR}{\emph{<chr>}}
\item{AGENCY_TYPE}{\emph{<chr>}}
\item{CESA}{\emph{<chr>}}
\item{COUNTY_CODE}{\emph{<chr>}}
\item{COUNTY}{\emph{<chr>}}
\item{CITY}{\emph{<chr>}}
\item{DISTRICT_CODE}{\emph{<chr>}}
\item{SCHOOL_CODE}{\emph{<chr>}}
\item{GRADE_GROUP}{\emph{<chr>}}
\item{CHOICE_IND}{\emph{<chr>}}
\item{DISTRICT_NAME}{\emph{<chr>}}
\item{SCHOOL_NAME}{\emph{<chr>}}
\item{LOW_GRADE}{\emph{<chr>}}
\item{HIGH_GRADE}{\emph{<dbl>}}
\item{GROUP_BY}{\emph{<chr>}}
\item{GROUP_BY_VALUE}{\emph{<chr>}}
\item{STUDENT_COUNT}{\emph{<int>}}
}
}
}
\source{
\url{https://dpi.wi.gov/wisedash/download-files/type?field_wisedash_upload_type_value=hs-completion-private-school}
}
\usage{
SPEC_FOR_PRIVATE_HS_COMPLETION
}
\description{
Column specification for data about Completion at Private High Schools
}
\keyword{datasets}
