% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SPEC_FOR_PS_ENROLLMENT}
\alias{SPEC_FOR_PS_ENROLLMENT}
\title{Column specification for data about Post-secondary Enrollment}
\format{
\subsection{SPEC_FOR_PS_ENROLLMENT}{

A list of 18 elements
\describe{
\item{SCHOOL_YEAR}{\verb{<chr>} the school year, in YYYY-YY format}
\item{AGENCY_TYPE}{\verb{<chr>} helps distinguish between district-run and independent charter schools}
\item{CHARTER_IND}{\verb{<chr>} whether or not a school is a charter}
\item{CESA}{\verb{<chr>} which CESA, if any, the school belongs to}
\item{COUNTY}{\verb{<chr>} the county that the school is located in}
\item{DISTRICT_CODE}{\verb{<chr>} the four-number code for each district, or "[Statewide]"}
\item{SCHOOL_CODE}{\verb{<chr>} the six-digit code for a school, or "[Districtwide]"}
\item{GRADE_GROUP}{\verb{<chr>} elementary, middle, high, k-12, etc.}
\item{DISTRICT_NAME}{\verb{<chr>} a human-readable name for a district. Not consistent across years}
\item{SCHOOL_NAME}{\verb{<chr>} a human-readable name for a school. Not consistent across years}
\item{INITIAL_ENROLLMENT}{\verb{<chr>} the time lag between graduation and enrollment}
\item{INSTITUTION_LOCATION}{\verb{<chr>} in-state or out-of-state}
\item{INSTITUTION_LEVEL}{\verb{<chr>} 2-year or 4-year}
\item{INSTITUTION_TYPE}{\verb{<chr>} public or private}
\item{GROUP_BY}{\verb{<chr>} demographic categories like race or gender}
\item{GROUP_BY_VALUE}{\verb{<chr>} demographic groups like Asian or Trans}
\item{GROUP_COUNT}{\verb{<int>} the number of students who could have enrolled in postsecondary education}
\item{STUDENT_COUNT}{\verb{<int>} the number of students who enrolled in postsecondary education}
}
}
}
\source{
\url{https://dpi.wi.gov/wisedash/download-files/type?field_wisedash_upload_type_value=Postsecondary}
}
\usage{
SPEC_FOR_PS_ENROLLMENT
}
\description{
Column specification for data about Post-secondary Enrollment
}
\keyword{datasets}
