% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SPEC_FOR_PUBLIC_HS_COMPLETION}
\alias{SPEC_FOR_PUBLIC_HS_COMPLETION}
\title{A specification for columns found only in reports about completing public high school.}
\format{
\subsection{SPEC_FOR_PUBLIC_HS_COMPLETION}{

A list of 5 elements
\describe{
\item{CHARTER_IND}{\verb{<chr>} whether or not a school is a charter}
\item{COHORT}{\verb{<int>} expected graduation year, e.g. 2010 for the 2009-10 school year}
\item{COMPLETION_STATUS}{\verb{<chr>} graduated normally, graduated with HSED, continuing, not}
\item{TIMEFRAME}{\verb{<chr>} whether it is the 4-, 5-, or 6-year rate, etc.}
\item{COHORT_COUNT}{\verb{<int>} the number of students who could have completed high school}
}
}
}
\source{
\url{https://dpi.wi.gov/wisedash/download-files/type?field_wisedash_upload_type_value=hs-completion}
}
\usage{
SPEC_FOR_PUBLIC_HS_COMPLETION
}
\description{
A specification for columns found only in reports about completing public high school.
}
\keyword{datasets}
