% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.R
\name{wrangle_forward}
\alias{wrangle_forward}
\title{Summarize forward results}
\usage{
wrangle_forward(.forward_data, .proficiency_threshold = 3L, ...)
}
\arguments{
\item{.forward_data}{a data frame with columns from \code{\link{SPEC_FOR_FORWARD}}}

\item{.proficiency_threshold}{\emph{<int?>} the lowest score code that counts as proficient (defaults is \code{3})}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> \emph{<chr>} grouping fields for computing testing and success rates.}
}
\value{
a data frame with \verb{len(...) + 6 columns}
\describe{
\item{...}{each grouping field}
\item{Students}{\emph{<int>} the total number of students in the group}
\item{Score}{\emph{<dbl>} the mean scale score of the group's students}
\item{Tested}{\emph{<int>} the number of students in the group who took the test}
\item{Proficient}{\emph{<int>} the number of students who were tested and achieved proficiency}
\item{Testing Rate}{\emph{<dbl>} the proportion of students in the group who took the test}
\item{Proficiency Rate}{\emph{<dbl>} the proportion of students who were tested and achieved proficiency}
}
}
\description{
Summarize forward results
}
