% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_completion.R
\name{wrangle_hs_completion}
\alias{wrangle_hs_completion}
\title{Convert raw data read from WISEDash Public files into a StriveTogether-friendly format}
\usage{
wrangle_hs_completion(.hs_data, .district = "4620")
}
\arguments{
\item{.hs_data}{a data frame in the format returned by \code{\link[=read_wisedash_public]{read_wisedash_public()}}}

\item{.district}{optional, the code for one district, defaults to Racine Unified's code.}
}
\value{
a tibble with five columns:
\describe{
\item{SCHOOL_YEAR}{\if{html}{\out{<chr>}} the school year in YYYY-YYYY format}
\item{Population}{\if{html}{\out{<chr>}} a demographic category defined by StriveTogether}
\item{Gender}{\if{html}{\out{<chr>}} a gender identity recognized by StriveTogether}
\item{Senior Class Size}{\if{html}{\out{<int>}} the number of students who could have graduated by the end of the school year}
\item{HS Completers}{\if{html}{\out{<int>}} the number of students who received a high school credential by the end of the school year}
}
}
\description{
Convert raw data read from WISEDash Public files into a StriveTogether-friendly format
}
\seealso{
\code{\link[=read_wisedash_public]{read_wisedash_public()}}
}
