% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_completion.R
\name{wrangle_ps_enrollment}
\alias{wrangle_ps_enrollment}
\title{Summarize post-secondary enrollment information}
\usage{
wrangle_ps_enrollment(.DATA, ...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> \emph{<chr>} grouping fields for computing post-secondary enrollment counts and rates}

\item{.ps_enrollment_data}{a data frame with columns from \code{\link{SPEC_FOR_PS_ENROLLMENT}}}
}
\value{
a data frame with \verb{length(...) + 6 columns}
\describe{
\item{...}{each grouping field}
\item{Students}{\emph{<int>} the total number of students in the group}
\item{Score}{\emph{<dbl>} the mean scale score of the group's students}
\item{Tested}{\emph{<int>} the number of students in the group who took the test}
\item{Proficient}{\emph{<int>} the number of students who were tested and achieved proficiency}
\item{Testing Rate}{\emph{<dbl>} the proportion of students in the group who took the test}
\item{Proficiency Rate}{\emph{<dbl>} the proportion of students who were tested and achieved proficiency}
}
}
\description{
Summarize post-secondary enrollment information
}
