% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_redactions.R
\name{wrangle_redactions}
\alias{wrangle_redactions}
\title{Handle situations where missing data at the school level has propagated to missing data at the district level}
\usage{
wrangle_redactions(.DATA, .groups, .sum_fields)
}
\arguments{
\item{.DATA}{\emph{<tbl>} data from WISEDash that includes both \code{SCHOOL_CODE} and \code{DISTRICT_CODE} fields.}

\item{.groups}{\emph{<chr>} a vector of column names that identify grouping variables}

\item{.sum_fields}{\emph{<chr>} a vector of column names that identify values to aggregate}
}
\value{
a data frame with the columns from \code{.groups} and \code{.sum_fields}
}
\description{
Handle situations where missing data at the school level has propagated to missing data at the district level
}
