% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SPEC_FOR_FORWARD}
\alias{SPEC_FOR_FORWARD}
\title{Column specification for data about performance on the Forward Exam}
\format{
\subsection{\code{SPEC_FOR_FORWARD}}{

A list of 20 elements
\describe{
\item{SCHOOL_YEAR}{<chr> the school year, in YYYY-YY format}
\item{AGENCY_TYPE}{<chr> helps distinguish between district-run and independent charter schools}
\item{CESA}{<chr>, which CESA, if any, the school belongs to}
\item{COUNTY}{<chr>, the county that the school is located in}
\item{DISTRICT_CODE}{<chr> the four-number code for each district, or "[Statewide]"}
\item{SCHOOL_CODE}{<chr> the six-digit code for a school, or "[Districtwide]"}
\item{GRADE_GROUP}{<chr> elementary, middle, high, k-12, etc.}
\item{CHARTER_IND}{<chr> whether or not a school is a charter}
\item{DISTRICT_NAME}{<chr> a human-readable name for a district. Not consistent across years}
\item{SCHOOL_NAME}{<chr> a human-readable name for a school. Not consistent across years}
\item{TEST_SUBJECT}{<chr> language or math}
\item{GRADE_LEVEL}{<chr> K-12}
\item{TEST_RESULT}{<chr> bad, mediocre, acceptable, awesome}
\item{TEST_RESULT_CODE}{<chr> one through four corresponding to \code{TEST_RESULT}}
\item{TEST_GROUP}{<chr> dunno}
\item{GROUP_BY}{<chr> demographic categories like race or gender}
\item{GROUP_BY_VALUE}{<chr> demographic groups like Asian or Trans}
\item{STUDENT_COUNT}{<int> the number of students with this test result}
\item{GROUP_COUNT}{<int> The number of students in this demographic category}
\item{FORWARD_AVERAGE_SCALE_SCORE}{<dbl> the mean Forward score in this group}
}
}
}
\source{
\url{https://dpi.wi.gov/wisedash/download-files/type?field_wisedash_upload_type_value=Forward}
}
\usage{
SPEC_FOR_FORWARD
}
\description{
Column specification for data about performance on the Forward Exam
}
\keyword{datasets}
