% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SPEC_FOR_FORWARD}
\alias{SPEC_FOR_FORWARD}
\title{Column specification for data about performance on the Forward Exam}
\format{
\subsection{\code{SPEC_FOR_FORWARD}}{

A list of 20 elements
\describe{
\item{SCHOOL_YEAR}{\emph{<chr>} the school year, in YYYY-YY format}
\item{AGENCY_TYPE}{\emph{<chr>} helps distinguish between district-run and independent charter schools}
\item{CESA}{\emph{<chr>} which CESA, if any, the school belongs to}
\item{COUNTY}{\emph{<chr>} the county that the school is located in}
\item{DISTRICT_CODE}{\emph{<chr>} the four-number code for each district, or "[Statewide]"}
\item{SCHOOL_CODE}{\emph{<chr>} the six-digit code for a school, or "[Districtwide]"}
\item{GRADE_GROUP}{\emph{<chr>} elementary, middle, high, k-12, etc.}
\item{CHARTER_IND}{\emph{<chr>} whether or not a school is a charter}
\item{DISTRICT_NAME}{\emph{<chr>} a human-readable name for a district. Not consistent across years}
\item{SCHOOL_NAME}{\emph{<chr>} a human-readable name for a school. Not consistent across years}
\item{TEST_SUBJECT}{\emph{<chr>} language or math}
\item{GRADE_LEVEL}{\emph{<chr>} K-12}
\item{TEST_RESULT}{\emph{<chr>} bad, mediocre, acceptable, awesome}
\item{TEST_RESULT_CODE}{\emph{<chr>} one through four corresponding to \code{TEST_RESULT}}
\item{TEST_GROUP}{\emph{<chr>} dunno}
\item{GROUP_BY}{\emph{<chr>} demographic categories like race or gender}
\item{GROUP_BY_VALUE}{\emph{<chr>} demographic groups like Asian or Trans}
\item{STUDENT_COUNT}{\emph{<int>} the number of students with this test result}
\item{GROUP_COUNT}{\emph{<int>} The number of students in this demographic category}
\item{FORWARD_AVERAGE_SCALE_SCORE}{\emph{<dbl>} the mean Forward score in this group}
}
}
}
\source{
\url{https://dpi.wi.gov/wisedash/download-files/type?field_wisedash_upload_type_value=Forward}
}
\usage{
SPEC_FOR_FORWARD
}
\description{
Column specification for data about performance on the Forward Exam
}
\keyword{datasets}
