% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SPEC_FOR_HS_COMPLETION}
\alias{SPEC_FOR_HS_COMPLETION}
\title{Column specification for data about High School Completion}
\format{
\subsection{SPEC_FOR_HS_COMPLETION}{

A list of 17 elements
\describe{
\item{SCHOOL_YEAR}{<chr> the school year, in YYYY-YY format}
\item{AGENCY_TYPE}{<chr> helps distinguish between district-run and independent charter schools}
\item{CHARTER_IND}{<chr> whether or not a school is a charter}
\item{CESA}{<chr>, which CESA, if any, the school belongs to}
\item{COUNTY}{<chr>, the county that the school is located in}
\item{DISTRICT_CODE}{<chr> the four-number code for each district, or "[Statewide]"}
\item{SCHOOL_CODE}{<chr> the six-digit code for a school, or "[Districtwide]"}
\item{GRADE_GROUP}{<chr> elementary, middle, high, k-12, etc.}
\item{DISTRICT_NAME}{<chr> a human-readable name for a district. Not consistent across years}
\item{SCHOOL_NAME}{<chr> a human-readable name for a school. Not consistent across years}
\item{COHORT}{<int> expected graduation year, e.g. 2010 for the 2009-10 school year}
\item{COMPLETION_STATUS}{<chr> graduated normally, graduated with HSED, continuing, not}
\item{GROUP_BY}{<chr> demographic categories like race or gender}
\item{GROUP_BY_VALUE}{<chr> demographic groups like Asian or Trans}
\item{TIMEFRAME}{<chr> whether it is the 4-, 5-, or 6-year rate, etc.}
\item{COHORT_COUNT}{<int> the number of students who could have completed high school}
\item{STUDENT_COUNT}{<int> the number of students who completed high school}
}
}
}
\source{
\url{https://dpi.wi.gov/wisedash/download-files/type?field_wisedash_upload_type_value=hs-completion}
}
\usage{
SPEC_FOR_HS_COMPLETION
}
\description{
Column specification for data about High School Completion
}
\keyword{datasets}
