% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SPEC_FOR_PRIVATE_HS_COMPLETION}
\alias{SPEC_FOR_PRIVATE_HS_COMPLETION}
\title{Column specification for data about Completion at Private High Schools}
\format{
\subsection{SPEC_FOR_PRIVATE_HS_COMPLETION}{

A list of 5 elements
\describe{
\item{COUNTY_CODE}{\emph{<chr>}}
\item{CITY}{\emph{<chr>}}
\item{CHOICE_IND}{\emph{<chr>}}
\item{LOW_GRADE}{\emph{<chr>}}
\item{HIGH_GRADE}{\emph{<chr>}}
}
}
}
\source{
\url{https://dpi.wi.gov/wisedash/download-files/type?field_wisedash_upload_type_value=hs-completion-private-school}
}
\usage{
SPEC_FOR_PRIVATE_HS_COMPLETION
}
\description{
Column specification for data about Completion at Private High Schools
}
\keyword{datasets}
