% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SPEC_FOR_PS_ENROLLMENT}
\alias{SPEC_FOR_PS_ENROLLMENT}
\title{Column specification for data about Post-secondary Enrollment}
\format{
\subsection{SPEC_FOR_PS_ENROLLMENT}{

\describe{
\item{SCHOOL_YEAR}{<chr> the school year, in YYYY-YY format}
\item{AGENCY_TYPE}{<chr> helps distinguish between district-run and independent charter schools}
\item{CHARTER_IND}{<chr> whether or not a school is a charter}
\item{CESA}{<chr>, which CESA, if any, the school belongs to}
\item{COUNTY}{<chr>, the county that the school is located in}
\item{DISTRICT_CODE}{<chr> the four-number code for each district, or "[Statewide]"}
\item{SCHOOL_CODE}{<chr> the six-digit code for a school, or "[Districtwide]"}
\item{GRADE_GROUP}{<chr> elementary, middle, high, k-12, etc.}
\item{DISTRICT_NAME}{<chr> a human-readable name for a district. Not consistent across years}
\item{SCHOOL_NAME}{<chr> a human-readable name for a school. Not consistent across years}
\item{INITIAL_ENROLLMENT}{<chr> the time lag between graduation and enrollment}
\item{INSTITUTION_LOCATION}{<chr> in-state or out-of-state}
\item{INSTITUTION_LEVEL}{<chr> 2-year or 4-year}
\item{INSTITUTION_TYPE}{<chr> public or private}
\item{GROUP_BY}{<chr> demographic categories like race or gender}
\item{GROUP_BY_VALUE}{<chr> demographic groups like Asian or Trans}
\item{GROUP_COUNT}{<int> the number of students who could have enrolled in postsecondary education}
\item{STUDENT_COUNT}{<int> the number of students who enrolled in postsecondary education}
}
}
}
\source{
\url{https://dpi.wi.gov/wisedash/download-files/type?field_wisedash_upload_type_value=Postsecondary}
}
\usage{
SPEC_FOR_PS_ENROLLMENT
}
\description{
Column specification for data about Post-secondary Enrollment
}
\keyword{datasets}
