% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_completion.R
\name{wrangle_hs_completion}
\alias{wrangle_hs_completion}
\title{Convert raw data read from WISEDash Public files into a StriveTogether-friendly format}
\usage{
wrangle_hs_completion(.DATA, ...)
}
\arguments{
\item{.DATA}{a data frame in the format returned by \code{\link[=read_wisedash_public]{read_wisedash_public()}}}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> \emph{<chr>} grouping fields for computing graduate counts and completion rates.}
}
\value{
a tibble with at \code{length(...) + 3} columns:
\describe{
\item{...}{the grouping variables specified by \code{...}}
\item{Students}{\emph{<int>} the number of students who could have graduated by the end of the school year}
\item{Graduates}{\emph{<int>} the number of students who received a high school credential by the end of the school year}
\item{Completion Rate}{\emph{<int>} the number of students who received a high school credential by the end of the school year}
}
}
\description{
Convert raw data read from WISEDash Public files into a StriveTogether-friendly format
}
\seealso{
\code{\link[=read_wisedash_public]{read_wisedash_public()}}
}
